remove legacy card checkout scripts and simplify the naming of existing ones
