+++ webpack/shared.js
@@ -11,7 +11,7 @@ module.exports = {
 		{
 			index: './client/index.js',
 			settings: './client/settings/index.js',
-			'blocks-checkout': './client/checkout/blocks/upe-split.js',
+			'blocks-checkout': './client/checkout/blocks/index.js',
 			woopay: './client/checkout/woopay/index.js',
 			'woopay-express-button':
 				'./client/checkout/woopay/express-button/index.js',
